Use (x)st as default terminal
`dst` is a custom version of xst with its name simply changed to dst.
This allows me to have separate configurations within my .Xresources
for each. I may repeat this and create executables for Weechat and
TaskWarrior to enable custom themes for each. Alternatively, I may try
implementing a command-line flag in xst for specifying the name to look
for in .Xresources